chore(desktop): split the move-workflow hook below the frontend size threshold
